create audio module in thetawave_interface
