+++ Cargo.toml
@@ -19,6 +19,7 @@ thiserror = "1.0"
 derive_more = "0.99.17"
 bevy_ecs_macros = "0.11.2"
 bevy_ecs = "0.11.2"
+bevy_kira_audio = { version = "0.16.0", features = ["mp3", "wav"] }
 
 [dependencies]
 bevy = { workspace = true }
@@ -38,7 +39,7 @@ strum_macros = { workspace = true }
 ron = "0.8.0"
 rand = "0.8.5"
 console_error_panic_hook = "0.1.7"
-bevy_kira_audio = { version = "0.16.0", features = ["mp3", "wav"] }
+bevy_kira_audio = { workspace = true }
 winit = "0.28.4"
 image = "0.24.6"
 thiserror = { workspace = true }
