+++ crates/thetawave_interface/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod audio;
 pub mod character_selection;
 pub mod game;
 pub mod health;
